+++ backend/src/auth/auth-controller.ts
@@ -1,4 +1,4 @@
-import User from '../models/user-model.js';
+import User from './user-model.js';
 import bcrypt from 'bcrypt';
 import { Request, Response } from 'express';
