+++ backend/src/auth/auth-router.ts
@@ -1,5 +1,5 @@
 import express from 'express';
-import {registerUser} from '../controllers/auth-controller.js';
+import {registerUser} from './auth-controller.js';
 
 const authRouter = express.Router();
